+++ src/server/cli.js
@@ -1,0 +1,17 @@
+/**
+ * src/server/cli.js: main executable script Express HTTP server
+ *
+ * Copyright 2016 Sudaraka Wijesinghe <sudaraka@sudaraka.org>
+ *
+ * This program comes with ABSOLUTELY NO WARRANTY
+ * This is free software, and you are welcome to redistribute it and/or modify
+ * it under the terms of the BSD 2-clause License. See the LICENSE file for more
+ * details.
+ */
+
+import express from 'express'
+
+const
+  app = express()
+
+app.listen(5000)